Move code from evaluator package to interpreter
