+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/CaptureP.scala
@@ -3,7 +3,7 @@ package com.rallyhealth.vapors.factfilter.dsl
 import cats.{Eval, Monoid}
 import com.rallyhealth.vapors.core.algebra.Expr
 import com.rallyhealth.vapors.core.data.TypedFact
-import com.rallyhealth.vapors.factfilter.evaluator.{ExprInput, ExprOutput}
+import com.rallyhealth.vapors.core.interpreter.{ExprInput, ExprOutput}
 
 import scala.annotation.implicitNotFound
