+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/ExprDsl.scala
@@ -4,10 +4,10 @@ import cats.data.NonEmptyList
 import cats.{Foldable, Id, Monoid}
 import com.rallyhealth.vapors.core.algebra.{ConditionBranch, Expr, ExprResult}
 import com.rallyhealth.vapors.core.data._
+import com.rallyhealth.vapors.core.interpreter.{ExprInput, InterpretExprAsResultFn}
 import com.rallyhealth.vapors.core.lens.NamedLens
 import com.rallyhealth.vapors.core.logic.{Conjunction, Disjunction, Negation}
 import com.rallyhealth.vapors.core.math.{Addition, Negative, Subtraction}
-import com.rallyhealth.vapors.factfilter.evaluator.{ExprInput, InterpretExprAsResultFn}
 
 object ExprDsl extends ExprBuilderSyntax with ExprBuilderCatsInstances with WrapExprSyntax {
